feat(gateway): allow sending messages to specific users
This allows clients to specify the `dmTo` field in messages. This should be the user ID of a connected user (if there are no matches, the message will be refused with an error). If this is specified, the message will not be saved to history, and it will be sent only to connections belonging to the specified user.